Rename KiwixErrorActivity to ErrorActivity and move it to error package
