+++ app/src/main/java/org/kiwix/kiwixmobile/utils/SplashActivity.java
@@ -22,7 +22,7 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 
-import org.kiwix.kiwixmobile.KiwixErrorActivity;
+import org.kiwix.kiwixmobile.error.ErrorActivity;
 import org.kiwix.kiwixmobile.main.MainActivity;
 
 import java.io.Serializable;
@@ -39,7 +39,7 @@ public class SplashActivity extends AppCompatActivity {
       @Override
       public void uncaughtException(Thread paramThread, Throwable paramThrowable) {
 
-        final Intent intent = new Intent(appContext, KiwixErrorActivity.class);
+        final Intent intent = new Intent(appContext, ErrorActivity.class);
 
         Bundle extras = new Bundle();
         extras.putSerializable("exception", (Serializable) paramThrowable);
